管理员可删除用户

+++ app/Http/Controllers/UsersController.php
@@ -11,7 +11,7 @@ class UsersController extends Controller
     public function __construct()
     {
         $this->middleware('auth',[
-           'except' => ['show', 'create', 'store']
+           'except' => ['show', 'create', 'store', 'index']
         ]);
         $this->middleware('guest', [
            'only' => 'create'
@@ -20,7 +20,8 @@ class UsersController extends Controller
 
     public function index()
     {
-        return view('users.index');
+        $users = User::paginate(10);
+        return view('users.index', compact('users'));
     }
     public function create()
     {
@@ -77,7 +78,9 @@ class UsersController extends Controller
 
     public function destroy(User $user)
     {
+        $this->authorize('destroy', $user);
         $user->delete();
+        session()->flash('success', '成功删除用户！');
         return redirect()->back();
     }
 }
